Update geometry edit types to simpler API;
use new isEmpty helper; extract lot selector to component; enforce type validation of feature collection;
