+++ app/components/project-geometries/project-area.js
@@ -47,12 +47,9 @@ export default class ProjectAreaComponent extends Component {
   projectAreaLayer = projectAreaLayer;
 
   @action
-  async save(finalGeometry) {
+  async save() {
     const model = this.get('model');
     const notifications = this.get('notificationMessages');
-    const { features: [{ geometry }] } = await finalGeometry;
-
-    model.set('projectArea', geometry);
 
     try {
       const savedProject = await model.save();
