+++ app/components/project-geometries/special-purpose-districts.js
@@ -2,6 +2,7 @@ import Component from '@ember/component';
 import { argument } from '@ember-decorators/argument';
 import { action } from '@ember-decorators/object';
 import { service } from '@ember-decorators/service';
+import isEmpty from '../../utils/is-empty';
 
 // Proposed Special Purpose Districts
 const proposedSpecialPurposeDistrictsLayer = {
@@ -47,7 +48,7 @@ export default class specialPurposeDistrictsComponent extends Component {
   init(...args) {
     super.init(...args);
 
-    if (!this.get('model.specialPurposeDistricts')) {
+    if (isEmpty(this.get('model.specialPurposeDistricts'))) {
       this.get('model').setDefaultSpecialPurposeDistricts();
     }
   }
